Excluindo a classe usuário e adicionando métodos no MedicoController

+++ src/main/java/br/com/teste/logus/agendaconsultorio/repository/MedicoRepository.java
@@ -3,5 +3,5 @@ package br.com.teste.logus.agendaconsultorio.repository;
 import br.com.teste.logus.agendaconsultorio.model.MedicoModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MedicoRepository extends CrudRepository<MedicoModel, Integer> {
+public interface MedicoRepository extends CrudRepository<MedicoModel, Long> {
 }
